Record_retv: run a new R session for every testcase

+++ Setup/genthat/record_retv.R
@@ -18,35 +18,46 @@ if (length(argv) != 1) {
   stop("Incorrect number of arguments")
 }
 
+globals <- new.env()
+globals$BENCH_DIR <- normalizePath(argv[[1]])
+globals$SOURCING_SEED <- integer(1234)
+globals$RUNNING_SEED <- integer(5689)
 
-BENCH_DIR <- normalizePath(argv[[1]])
-SOURCING_SEED <- integer(1234)
-RUNNING_SEED <- integer(5689)
+
+source_reproducible <- function(globals, Rfile) {
+  .Random.seed <<- globals$SOURCING_SEED
+  source(Rfile)
+  .Random.seed <<- globals$RUNNING_SEED
+  function_to_run()
+}
+
+
 
 
 # running the scripts might create files in the working directory
 setwd(tempdir())
 
-for (Rfile in list.files(BENCH_DIR, pattern = "*.R", full.names = TRUE, recursive = TRUE)) {
+for (Rfile in list.files(globals$BENCH_DIR, pattern = ".*R$", full.names = TRUE, recursive = TRUE)) {
   extfile <- gsub(".R$", ".ext", Rfile)
   message("Processing ", Rfile, " to generate ", extfile)
 
-  retv <- NULL
-  tryCatch(
-    capture.output({
-      .Random.seed <<- SOURCING_SEED
-      source(Rfile)
-      .Random.seed <<- RUNNING_SEED
-      retv <- function_to_run()
+  if (file.exists(extfile)) {
+    message("SKIP: Ext file exists ", extfile)
+    next
+  }
 
+  tryCatch(
+    {
+      # Source the file and call the function in a new R session
+      retv <- callr::r(source_reproducible, args=list(globals, Rfile))
 
       l <- list(
-        sourcing_seed=SOURCING_SEED,
-        running_seed=RUNNING_SEED,
+        sourcing_seed=globals$SOURCING_SEED,
+        running_seed=globals$RUNNING_SEED,
         retv=retv
       )
       saveRDS(l, file=extfile)
-    }),
+    },
     error=function(e) {
       message("Failed to run ", Rfile, ": ", e)
       failname <- paste0(Rfile, ".failing")
